Organised all the new dropdown navigation code in dropdown-navigation directory and files
